requirements.txt
boto3==1.26.151
botocore==1.29.151
gspread==5.9.0
oauth2client==4.1.3
pandas==1.4.2
python-dotenv==1.0.0
